// src/Convience.Web/Managent/src/app/pages/system-manage/model/IndicatorDetail.ts
export class IndicatorDetail {
  paramid?: string;
  department?: string;
  nickname?: string;
  parentid?: string;
  fomular?: string;
  unit?: string;
  weight?: number;
  goalyear?: string;
  sort?: number;
  enable?: boolean;
  createid?: string;
  createtime?: Date;
  modifyid?: string;
  modifytime?: Date;
}